pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gai-cli"
version = "0.1.0"
description = "AI-powered git assistant for commit messages, merge reviews, and intelligent diff analysis"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "AGPL-3.0-or-later"}
authors = [
    {name = "gai contributors"}
]
keywords = ["git", "ai", "commit", "llm", "groq", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "click>=8.1.0",
    "requests>=2.32.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
postgresql = ["psycopg2-binary>=2.9.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/gai"
Documentation = "https://github.com/yourusername/gai#readme"
Repository = "https://github.com/yourusername/gai.git"
Issues = "https://github.com/yourusername/gai/issues"

[project.scripts]
gai = "gai.gai:cli"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["gai", "gai.core", "gai.ai", "gai.commands", "gai.ui", "gai.utils"]

[tool.setuptools.package-data]
gai = ["*.yaml", "*.yml"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=gai",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
]

[tool.coverage.run]
source = ["src/gai"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/.pytest_cache/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
